refactor(form): consume da-sc-sdk from npm via shared nx/deps/
