--- nx/deps/da-form/src/index.js
@@ -1,5 +1,0 @@
-import { Validator } from '@cfworker/json-schema';
-import { fromHtmlIsomorphic } from 'hast-util-from-html-isomorphic';
-import { selectAll } from 'hast-util-select';
-
-export { Validator, fromHtmlIsomorphic, selectAll };
