+++ nx/deps/da-sc-sdk/src/index.js
@@ -1,0 +1,3 @@
+import { createEngine, convertJsonToHtml, convertHtmlToJson } from '@adobe/da-sc-sdk';
+
+export { createEngine, convertJsonToHtml, convertHtmlToJson };
