+++ package.json
@@ -18,6 +18,7 @@
     "test:file:watch": "wtr --config ./web-test-runner.config.mjs --node-resolve --port=2000 --coverage --watch",
     "build:mdast": "esbuild --format=esm --minify ./nx/deps/mdast/src/index.js --bundle --outfile=./nx/deps/mdast/dist/index.js",
     "build:codemirror": "esbuild --format=esm --minify ./nx/deps/codemirror/src/index.js --bundle --outfile=./nx/deps/codemirror/dist/index.js",
+    "build:da-sc-sdk": "esbuild --format=esm --minify ./nx/deps/da-sc-sdk/src/index.js --bundle --outfile=./nx/deps/da-sc-sdk/dist/index.js",
     "build:da-lit": "esbuild --format=esm --minify ./deps/lit/src/index.js --bundle --outfile=./deps/lit/dist/index.js",
     "build:ml-lit": "esbuild --format=esm --minify ./nx/deps/ml-lit/src/index.js --bundle --outfile=./nx/deps/ml-lit/dist/index.js",
     "build:virtualizer": "esbuild --format=esm --minify ./nx/deps/virtualizer/src/index.js --bundle --outfile=./nx/deps/virtualizer/dist/index.js",
@@ -62,18 +63,16 @@
     "stylelint-config-standard": "40.0.0"
   },
   "dependencies": {
+    "@adobe/da-sc-sdk": "^0.1.0",
     "@adobe/helix-markdown-support": "^7.1.18",
     "@adobe/mdast-util-gridtables": "^4.0.6",
     "@adobe/remark-gridtables": "^3.0.6",
-    "@cfworker/json-schema": "^4.1.1",
     "@codemirror/lang-json": "^6.0.2",
     "@fsegurai/codemirror-theme-github-light": "^6.2.2",
     "@lit-labs/virtualizer": "^2.0.16",
     "codemirror": "^6.0.2",
     "esbuild": ">=0.25.0",
-    "hast-util-from-html-isomorphic": "^2.0.0",
     "hast-util-raw": "^9.0.4",
-    "hast-util-select": "^6.0.4",
     "hast-util-to-html": "^9.0.3",
     "lit": "^3.2.0",
     "mdast-util-to-hast": "^13.2.0",
